Went through and updated Unlinked Mentions for all Spells where appropriate
Ignoring all unrelated, such as references to the word "Light" which did not refer to the spell explicitly.

NB: Because there is a spell called "Divination", have tried to come up with a usable system to differentiate spells which refer:

A) Explicitly to the Divination spell
B) Spells which are classified as from the Divination School of Magic
C) spells which refer to the Divination School, but are not Divination themselves

Specifically important for C, because you may find spells which are eg. Abjuration magic that protects from divination magic. If we used one [[]] for both, those abjuration spells would be grouped in with the others.

Actually changing the tag to refer to [[Divination Spells]] will be done later. Hopefully by adding the list to the top it's still quite easy to cross-reference.